+++ lib/feature/transactions/repository/machine_transaction_repo.dart
@@ -1,3 +1,5 @@
+import 'dart:convert';
+
 import 'package:browny_applications_new/core/data/repo/app_repository.dart';
 import 'package:browny_applications_new/core/data/remote/models/request/machine_order_request.dart';
 import 'package:browny_applications_new/core/data/remote/models/request/machine_order_review_request.dart';
@@ -10,6 +12,7 @@ import 'package:browny_applications_new/core/utils/app_extensions.dart';
 import 'package:browny_applications_new/core/utils/repo_result.dart';
 import 'package:browny_applications_new/feature/authentication/error/authen_exception.dart';
 import 'package:dio/dio.dart';
+import 'package:flutter/foundation.dart';
 
 mixin MachineTransactionDataSourceMixin {
   /// API fetch รายละเอียดเครื่อง (machine detail)
@@ -46,6 +49,47 @@ class MachineRepo extends AppRepository with MachineTransactionDataSourceMixin {
     String machineId,
   ) async {
     try {
+      // if (kDebugMode) {
+      //   return RepoResult.success(
+      //     data: MachineDetailResponse.fromJson(
+      //       jsonDecode('''
+      // {
+      //     "id": 13,
+      //     "store_name": {
+      //         "th": "ตลาดคูล - บางกรวย",
+      //         "en": "Cool Market - Bang Kruai",
+      //         "zh": "-"
+      //     },
+      //     "status": "Busy",
+      //     "finish_datatime": "16:30",
+      //     "remaining_time": "00:00:50",
+      //     "machine_no": "1",
+      //     "machine_image": "https://dev.abgroup.co.th/storage/galleries/s3nbAR7QLSPpZ9aSTWSyGV9nXQZy6JhsPgVvaJKL.png",
+      //     "name": {
+      //         "th": "เครื่องซัก 1 - 16.00 กก.",
+      //         "en": "Washer 1 - 16.00 kg",
+      //         "zh": "洗衣机 1 - 16.00 公斤"
+      //     },
+      //     "order_id": "BBL11234",
+      //     "receipt_no": "1234",
+      //     "startTime": "2026-03-05 16:00:23",
+      //     "addTime": [],
+      //     "program_image": "https://dev.abgroup.co.th/storage/galleries/fFr4U1j8jv8NYy1adbWjJvsuT9BaX9AOWAnoOASF.png",
+      //     "program_name": {
+      //     "th": "น้ำเย็น",
+      //     "en": "Cold Water",
+      //     "zh": "冷水"
+      //     },
+      //     "machine_type": {
+      //         "th": "เครื่องอบ",
+      //         "en": "Washer",
+      //         "zh": "烘干机"
+      //     }
+      // }
+      // '''),
+      //     ),
+      //   );
+      // }
       final response = await requireRemote.fetchMachineDetail(machineId);
       if (!response.isSuccessful) {
         return RepoResult.empty();